requirements.txt
opencv-python
pydicom==2.4.4
tifffile==2024.5.22
clip==0.2.0
numpy==2.0.0
openai_clip==1.0.1
torch-fidelity
pillow
Pillow==10.3.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.64.1
setuptools==69.5.1 # setuptools最新バージョンのエラー回避のため
matplotlib==3.9.0
scikit-learn==1.5.0
scipy==1.13.1
jupyter==1.0.0
pandas==2.2.2
japanize-matplotlib==1.1.3
wandb==0.18.6
plotly==5.24.1
openpyxl==3.1.5
seaborn==0.13.2
scikit-image==0.22.0
gryds
dominate==2.9.1
visdom==0.2.4
POT==0.9.6.post1